app/backend/requirements.txt
fastapi>=0.111.0,<1.0
uvicorn>=0.29.0,<1.0
biopython>=1.83,<2.0
pyfaidx>=0.8.1,<1.0
bamnostic>=1.3,<2.0
python-multipart>=0.0.9,<1.0
